requirements.txt
numpy==1.21.2
Pillow==8.3.2
pandas==1.3.3
progressbar2==3.53.1
matplotlib==3.4.3
